typedef Vector3f -> Vector3d

+++ capturability/include/base.h
@@ -9,7 +9,7 @@ namespace Capt {
 
 typedef Eigen::Vector2f vec2_t;
 typedef Eigen::Vector2i vec2i_t;
-typedef Eigen::Vector3f vec3_t;
+typedef Eigen::Vector3d vec3_t;
 typedef std::vector<vec2_t> arr2_t;
 typedef std::vector<vec3_t> arr3_t;
